// 06032026/sce-front/src/app/model/reportes/filtroReporteMesasPorEstadoMesa.ts
export class FiltroReporteMesasPorEstado {
    esquema: string;
    proceso: string;
    idEleccion:number;
    codigoEleccion: string;
    idProceso:number;
    eleccion: string;
    tipoReporte: string;
    codigoAmbitoElectoral: string;
    ambitoElectoral: string;
    codigoCentroComputo: string;
    centroComputo: string;
    codigoEstadoMesa: string;
    estadoMesa: string;
    idEstadoMesa: number;
    ubigeoNivelUno: string;
    ubigeoNivelDos: string;
    ubigeoNivelTres: string;
    porDistrito: boolean;
}